pyproject.toml
[project]
name = "llama3-jax"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "cloudpathlib>=0.20.0",
  "hydra-core>=1.3.2",
  "jax>=0.5.0",
  "jaxlib>=0.5.0",
  "loguru>=0.7.3",
  "rich>=13.9.4",
  "tiktoken>=0.9.0",
]

[tool.uv]
package = true # This makes uv automatically add deepchain_mrna to PYTHONPATH
